#35: refactoring of base classes structure
----
Decouple Action base classes for a better flexibility
https://github.com/php-kitchen/yii2-domain/issues/35
